fix(test_harness): build dist/ once via vitest globalSetup, not per-harness
Previously, every `bootIntegrationHarness` call invoked an mtime-then-
maybe-build sequence inline. Under vitest's default parallel pool, multiple
workers could race on the build — and `npm run build`'s `prebuild` step
runs `rm -rf dist && eslint`, which deletes dist/ from under any worker
mid-spawn or mid-load of the dist Worker file. Failure modes ranged from
ERR_MODULE_NOT_FOUND on a half-deleted dist tree to vacuous passes when
one worker finished a build during another worker's mtime check.

Move the build to vitest's globalSetup hook (src/test_harness/global_setup.ts).
globalSetup runs once in the main process before any worker thread spawns,
so dist/ is stable for the entire test run; no worker writes to dist/. The
mtime-skip is preserved so unit-only `npx vitest run` invocations don't
pay the build cost when dist/ is already fresh.

The harness's old `ensureDistBuilt()` becomes `resolveDistWorkerUrl()` —
a read-only resolver that throws with a pointer to globalSetup if the
file is missing (the caller-outside-vitest case).

Verified: cold-run with `rm -rf dist` rebuilds via globalSetup, full suite
green. Five consecutive full-suite runs all 729/729 — no flakes.
